Remove unecessary assignment and conditional with :reset and nil
I think all the harassment of setting prev_word and next_word to nil
or to :reset is related with trying to iterate from 0 indexed.

I think we can just jump and get the first 2 elements and compare them.

words[word_ix] is the first
words[word_ix+1] is the second

And at iteration they guard/maintain this relation.