Remove break
We really don't need to break.
We can just go on the iteration and keep on swaping
the words till the end of the array.

Then come back and do it again and again till
we pass it without swapping any.

This is the bubble way of doing it.
We just have to have a way of signaling that
we passed the iteration without making any swaps.
We do this by setting more to -1 everytime we swap some word.
At the end we do more += 1. If more begins with -1 it will turn
to 0 and we go on iterating. If it's 0 already, it'll turn to 1
and we can stop as the array is sorted.